requirements.txt
discord.py>=2.0,<3.0
python-dotenv>=1.0,<2.0
requests>=2.20,<3.0
